[C] Cast precision expression for ‘%.*s’ to ‘int’.
Omission from 2010-05-18,
"Don't mutate string in ‘checkidentifier’".
* src/rcsmap.c (checkidentifier):
Cast precision expression for ‘%.*s’ to ‘int’.

+++ src/rcsmap.c
@@ -105,7 +105,7 @@ checkidentifier (char const *id, int delimiter, register bool dotok)
              && c != delim)
         id++;
       PFATAL ("invalid %s `%.*s'", dotok ? "identifier" : "symbol",
-              id - temp, temp);
+              (int) (id - temp), temp);
     }
   return id;
 }
